--- app/components/Todo.js
@@ -1,4 +1,0 @@
-import React from 'react';
-export default (todo) => (<div>
-    {JSON.stringify(todo)}
-</div>);
