+++ app/components/todos/AddTodo.js
@@ -1,7 +1,11 @@
 import React, {Component} from 'react';
-import todoActions from '../actions/todo';
+import todoActions from '../../actions/todo';
+
+import styles from './todos.scss';
+
 export default class AddTodo extends Component {
   render() {
-    return (<button onClick={() => todoActions.addTodo('SOME TODO')}>Add Todo</button>);
+    return (<button onClick={() => todoActions.addTodo('SOME TODO')}
+                    className={styles.TodoButton}>Add Todo</button>);
   }
 }
