+++ app/components/todos/Todo.js
@@ -1,0 +1,4 @@
+import React from 'react';
+export default (todo) => (<div>
+    {JSON.stringify(todo)}
+</div>);
