+++ app/components/todos/todos.scss
@@ -1,6 +1,5 @@
-
 @import '../variables.scss';
 
 .TodoButton {
-    padding: $padding-base-vertical $padding-base-horizontal;
+  padding: $padding-base-vertical $padding-base-horizontal;
 }
